devtools/server/tests/mochitest/test_inspector-release.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gWalker = null;
let gClient = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

addTest(function setup() {
  const url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    const {InspectorFront} = require("devtools/shared/fronts/inspector");
    const inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testReleaseSubtree() {
  let originalOwnershipSize = 0;
  let longlist = null;
  let firstChild = null;
  promiseDone(gWalker.querySelectorAll(gWalker.rootNode, "#longlist div").then(list => {
    // Make sure we have the 26 children of longlist in our ownership tree.
    is(list.length, 26, "Expect 26 div children.");
    // Make sure we've read in all those children and incorporated them
    // in our ownership tree.
    return list.items();
  }).then((items)=> {
    originalOwnershipSize = assertOwnership();

    // Here is how the ownership tree is summed up:
    // #document                      1
    //   <html>                       1
    //     <body>                     1
    //       <div id=longlist>        1
    //         <div id=a>a</div>   26*2 (each child plus it's singleTextChild)
    //         ...
    //         <div id=z>z</div>
    //                             -----
    //                               56
    is(originalOwnershipSize, 56, "Correct number of items in ownership tree");
    firstChild = items[0].actorID;
  }).then(() => {
    // Now get the longlist and release it from the ownership tree.
    return gWalker.querySelector(gWalker.rootNode, "#longlist");
  }).then(node => {
    longlist = node.actorID;
    return gWalker.releaseNode(node);
  }).then(() => {
    // Our ownership size should now be 53 fewer
    // (we forgot about #longlist + 26 children + 26 singleTextChild nodes)
    const newOwnershipSize = assertOwnership();
    is(newOwnershipSize, originalOwnershipSize - 53,
      "Ownership tree should be lower");
    // Now verify that some nodes have gone away
    return checkMissing(gClient, longlist);
  }).then(() => {
    return checkMissing(gClient, firstChild);
  }).then(runNextTest));
});

addTest(function cleanup() {
  gWalker = null;
  gClient = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
